scraper/requirements.txt
requests>=2.25.0
beautifulsoup4>=4.9.0
lxml>=4.6.0
sqlalchemy>=1.4.20,<2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=0.20.0
numpy>=1.26.0
pandas>=2.0.0
